requirements.txt
glfw==2.5.6
numpy==1.24.2
pyglet==2.0.4
PyOpenGL==3.1.6
PyGLM==2.7.0
pillow==9.4.0
imgui==1.4.1
jsonpickle==3.0.1